refactor(matches): split useMatchesPageLogic into focused sub-hooks
Extract data fetching, single-match CRUD, and bulk operations into
dedicated hooks (useMatchesData, useMatchesCrudActions, useMatchesBulkActions).
useMatchesPageLogic becomes a thin orchestrator wiring the sub-hooks together.
