+++ src/app/admin/matches/page.tsx
@@ -1,4 +1,4 @@
-import {AdminMatchesContainer} from '@/features/matches/components';
+import {AdminMatchesContainer} from '@/features/matches';
 
 export default function MatchesAdminPage() {
   return <AdminMatchesContainer />;
