+++ src/features/matches/hooks/index.ts
@@ -1,4 +1,7 @@
+export {useMatchesBulkActions} from './useMatchesBulkActions';
+export {useMatchesCrudActions} from './useMatchesCrudActions';
+export {useMatchesData} from './useMatchesData';
 export {useMatchesFilters} from './useMatchesFilters';
 export {useMatchesModals} from './useMatchesModals';
-export {useMatchesStandings} from './useMatchesStandings';
 export {useMatchesPageLogic} from './useMatchesPageLogic';
+export {useMatchesStandings} from './useMatchesStandings';
